44 create login page

+++ ScheduleLNU.Presentation/Areas/Settings/Controllers/EventStyleSettingsController.cs
@@ -1,18 +1,18 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using ScheduleLNU.BusinessLogic.DTOs;
+using ScheduleLNU.BusinessLogic.Extensions;
 using ScheduleLNU.BusinessLogic.Services.Interfaces;
-using ScheduleLNU.DataAccess.Entities;
 
 namespace ScheduleLNU.Presentation.Areas.Settings.Controllers
 {
     [Area("settings")]
     [Route("[area]/event-styles")]
-
-    // TODO: move logs to service
+    [Authorize]
     public class EventStyleSettingsController : Controller
     {
         private readonly ILogger<EventStyleSettingsController> logger;
@@ -29,33 +29,21 @@ namespace ScheduleLNU.Presentation.Areas.Settings.Controllers
         [HttpGet]
         public async Task<IActionResult> EventStyles()
         {
-            // TODO: remove logs of user routing
-            logger.LogInformation("Student oppened event styles page");
-            const int studentId = 228;
+            var studentId = HttpContext.GetStudentId();
+            if (studentId is null)
+            {
+                return StatusCode(401);
+            }
 
             IEnumerable<EventStyleDto> eventStyles = await eventStyleService.GetAllAsync(studentId);
-
-            logger.LogInformation("Student viwed all event styles {Lenght}", eventStyles.Count());
-
             return View(eventStyles);
         }
 
         [HttpPost]
         [Route("edit")]
 
-        // TODO: parameters as EventStyleDTO
-        public IActionResult EventStyleEdit(int styleId, int studentId, string foreColor, string backColor, string title)
+        public IActionResult EventStyleEdit(EventStyleDto eventStyleDto)
         {
-            var eventStyleDto = new EventStyleDto
-            {
-                Id = styleId,
-                StudentId = studentId,
-                ForeColor = foreColor,
-                BackColor = backColor,
-                Title = title
-            };
-
-            logger.LogInformation("Student tries to edit event style {styleId}", styleId);
             return View("EventStyleEdit", eventStyleDto);
         }
 
@@ -66,43 +54,35 @@ namespace ScheduleLNU.Presentation.Areas.Settings.Controllers
             await eventStyleService.EditAsync(eventStyleDto);
             if (ModelState.IsValid)
             {
-                logger.LogInformation("Student updated event style {eventStyleId}: title - {eventStyleTitle}," +
-                    "fore color - {eventStyleForeColor} and back color - {eventStyleBackColor} to the list of event styles",
-                    eventStyleDto.Id, eventStyleDto.Title, eventStyleDto.ForeColor, eventStyleDto.BackColor);
-
                 // TODO: to nameof of action
                 return RedirectToAction("EventStyles", new { studentId = eventStyleDto.StudentId });
             }
 
-            logger.LogInformation("Student failed to edit event style");
             return new StatusCodeResult(500);
         }
 
         [HttpGet]
         [Route("event-style-preview")]
-        public IActionResult EventStylePreview(int studentId)
+        public IActionResult EventStylePreview()
         {
-            logger.LogInformation("Student {studentID} creates new event style", studentId);
-
-            // TODO: Remove first parameter
-            return View("EventStylePreview", new EventStyleDto());
+            return View(new EventStyleDto());
         }
 
         [HttpPost]
         [Route("add-style")]
         public async Task<IActionResult> AddStyle(EventStyleDto eventStyleDto)
         {
-            eventStyleDto.StudentId = 228;
+            var studentId = HttpContext.GetStudentId();
+            if (studentId is null)
+            {
+                return StatusCode(401);
+            }
+
             await eventStyleService.AddAsync(eventStyleDto);
             if (ModelState.IsValid)
             {
-                logger.LogInformation("Student added new event style with a title - {eventStyleTitle}," +
-                    "fore color - {eventStyleForeColor} and back color - {eventStyleBackColor}" +
-                    " to the list of event styles", eventStyleDto.Title, eventStyleDto.ForeColor, eventStyleDto.BackColor);
                 return RedirectToAction("EventStyles", new { studentId = eventStyleDto.StudentId });
             }
-
-            logger.LogInformation("Student failed to add new event style");
 
             // TODO: use propeper status code
             return new StatusCodeResult(500);
@@ -119,18 +99,18 @@ namespace ScheduleLNU.Presentation.Areas.Settings.Controllers
         [Route("delete")]
         public async Task<IActionResult> DeleteStyle(int styleId)
         {
-            logger.LogInformation($"Deleting style {styleId}");
-            int studentId = 228;
+            var studentId = HttpContext.GetStudentId();
+            if (studentId is null)
+            {
+                return StatusCode(401);
+            }
+
             await eventStyleService.DeleteAsync(studentId, styleId);
             if (ModelState.IsValid)
             {
-                logger.LogInformation($"Successfully deleted {styleId}");
-
                 // TODO: Use proper status code
                 return new StatusCodeResult(201);
             }
-
-            logger.LogInformation("Student failed to delete style");
 
             // TODO: Use proper status code
             return new StatusCodeResult(500);
